+++ monitoring/monitoring/dashboards.go
@@ -1,7 +1,6 @@
 package monitoring
 
 import (
-	"fmt"
 	"strings"
 	"unicode"
 
@@ -52,109 +51,6 @@ const (
 	// BitsPerSecond, e.g. for representing network and disk IO.
 	BitsPerSecond UnitType = "bps"
 )
-
-// ObservablePanelOptions declares options for visualizing an Observable.
-type ObservablePanelOptions struct {
-	options                []ObservablePanelOption
-	disableAlertThresholds bool
-
-	// parameters used by other parts of the generator
-	unitType UnitType
-}
-
-// PanelOptions provides a builder for customizing an Observable visualization, starting
-// with recommended defaults.
-func PanelOptions() ObservablePanelOptions {
-	return ObservablePanelOptions{
-		options: []ObservablePanelOption{
-			optionMinimal(),
-			optionWithThresholds(),
-			optionWithUnits(Number),
-			// misc defaults
-			func(o Observable, g *sdk.GraphPanel) {
-				g.Targets[0].LegendFormat = "value"
-				g.Yaxes[0].Min = sdk.NewFloatString(0.0)
-			},
-		},
-	}
-}
-
-// MinimalPanelOptions provides a builder for customizing an Observable visualization
-// starting with an extremely minimal graph panel.
-func MinimalPanelOptions() ObservablePanelOptions {
-	return ObservablePanelOptions{
-		options: []ObservablePanelOption{
-			optionMinimal(),
-		},
-	}
-}
-
-// Min sets the minimum value of the Y axis on the panel. The default is zero.
-func (p ObservablePanelOptions) Min(min float64) ObservablePanelOptions {
-	p.options = append(p.options, func(o Observable, g *sdk.GraphPanel) {
-		g.Yaxes[0].Min = sdk.NewFloatString(min)
-	})
-	return p
-}
-
-// Min sets the minimum value of the Y axis on the panel to auto, instead of
-// the default zero.
-//
-// This is generally only useful if trying to show negative numbers.
-func (p ObservablePanelOptions) MinAuto() ObservablePanelOptions {
-	p.options = append(p.options, func(o Observable, g *sdk.GraphPanel) {
-		g.Yaxes[0].Min = nil
-	})
-	return p
-}
-
-// Max sets the maximum value of the Y axis on the panel. The default is auto.
-func (p ObservablePanelOptions) Max(max float64) ObservablePanelOptions {
-	p.options = append(p.options, func(o Observable, g *sdk.GraphPanel) {
-		g.Yaxes[0].Max = sdk.NewFloatString(max)
-	})
-	return p
-}
-
-// LegendFormat sets the panel's legend format, which may use Go template strings to select
-// labels from the Prometheus query.
-func (p ObservablePanelOptions) LegendFormat(format string) ObservablePanelOptions {
-	p.options = append(p.options, func(o Observable, g *sdk.GraphPanel) {
-		g.Targets[0].LegendFormat = format
-	})
-	return p
-}
-
-// Unit sets the panel's Y axis unit type.
-func (p ObservablePanelOptions) Unit(t UnitType) ObservablePanelOptions {
-	p.unitType = t
-	p.options = append(p.options, optionWithUnits(t))
-	return p
-}
-
-// Interval declares the panel's interval in milliseconds.
-func (p ObservablePanelOptions) Interval(ms int) ObservablePanelOptions {
-	p.options = append(p.options, func(o Observable, g *sdk.GraphPanel) {
-		g.Targets[0].Interval = fmt.Sprintf("%dms", ms)
-	})
-	return p
-}
-
-func (p ObservablePanelOptions) Thresholds(enabled bool) {
-	p.disableAlertThresholds = !enabled
-}
-
-// WithOption renders an additional option.
-func (p ObservablePanelOptions) WithOption(op ObservablePanelOption) ObservablePanelOptions {
-	p.options = append(p.options, op)
-	return p
-}
-
-func (p ObservablePanelOptions) generatePipeline(o Observable) []ObservablePanelOption {
-	if p.disableAlertThresholds {
-		return p.options
-	}
-}
 
 // setPanelSize is a helper to set a panel's size.
 func setPanelSize(p *sdk.Panel, width, height int) {
